Refactors Actions Into Their Own Files
